fix(types): enable reportArgumentType gate; clear real findings (49→0)
Promote reportArgumentType to error and clear every real finding, so the
pyright gate now covers the wrong-VALUE class (None into str, float into int,
list-typed record where a dict was meant) on top of wrong-kwarg/wrong-arg.

Real bugs fixed (not just typing noise):
- scheduler.py: get_verified_hosts() returns list[str], but the verification
  filter did h["host_id"] on each id string — a guaranteed
  "string indices must be integers" crash whenever it ran. Use
  set(ve.get_verified_hosts()).
- routes/marketplace.py: search passed `body.min_vram_gb or None` /
  `max_price_cents or None` into params used as `if x > 0` → `None > 0`
  TypeError when the filter was omitted. Drop `or None` (engine already uses
  0/"" sentinels); widen vram params to float to match the API model.
- volumes.py / worker_agent.py: implicit-Optional defaults
  (instance_id: str = None) corrected to `str | None`; coerce the resolved
  instance_id from the row so the unmount call stays typed.
- routes/mfa.py: guard auth_data.credential_data before bytes(cred_data)
  (would have been bytes(None)).

Clean type corrections:
- bitcoin._wallet_rpc_call -> Any (JSON-RPC returns arbitrary JSON).
- privacy._cascade_delete: coerce str category -> DataCategory enum key.
- jurisdiction.JurisdictionConstraint.trust_tier -> Optional[str] (callers
  pass None to mean "no constraint").
- host_profiles._quote_cmd accepts list[Any] (body already str()-coerces).
- routes/health checks dict typed dict[str, Any]; artifacts _CACHE_AUTO: Any.
- Library-boundary casts: stripe business_type, terminal paramiko
  Channel/socket + memoryview sends.

Gate redesign: scripts/pyright_gate.py now enforces reportCallIssue AND
reportArgumentType at zero tolerance, filtering only the psycopg dynamic-SQL
LiteralString noise (23 "QueryNoTemplate" findings — parameterising that
pattern is a separate initiative). reportCallIssue 0, reportArgumentType real
findings 0.

+++ marketplace.py
@@ -58,7 +58,7 @@ class MarketplaceEngine:
         provider_id: str,
         host_id: str,
         gpu_model: str,
-        vram_gb: int,
+        vram_gb: float,
         ask_cents_per_hour: int,
         gpu_count_total: int = 1,
         region: str = "",
@@ -145,7 +145,7 @@ class MarketplaceEngine:
     def search_offers(
         self,
         gpu_model: str = "",
-        min_vram_gb: int = 0,
+        min_vram_gb: float = 0,
         max_price_cents: int = 0,
         region: str = "",
         canada_only: bool = False,
